ver 0.2.0 - ignore feeds without md5, implement azcopy module, add azcopy env variables, add preserve src feeds env variable

+++ src/ubs_landing_zone/pipeline.py
@@ -1,10 +1,12 @@
 import os
 import shutil
 import tarfile
+from datetime import datetime
 from pathlib import Path
 import hashlib
 import time
 import tempfile
+import subprocess
 
 from .az_copy import AzCopy
 from loguru import logger
@@ -16,13 +18,15 @@ class Pipeline:
         checksum_extension: str, 
         algorithm: str, 
         failed_dir: Path,
-        processing_dir: Path
+        processing_dir: Path,
+        preserve_source_feeds: bool = False
     ):
         self._az_copy: AzCopy = az_copy
         self._checksum_extension: str = checksum_extension
         self._algorithm: str = algorithm
         self._failed_dir: Path = failed_dir
         self._processing_dir: Path = processing_dir
+        self._preserve_source_feeds: bool = preserve_source_feeds
 
     def run(self, feed: Path) -> None:
         start_time = time.time()
@@ -36,16 +40,17 @@ class Pipeline:
             ordered_feed_content: list[Path] = self._order_feed_content(unpacked_dir, feed)
             self._upload(ordered_feed_content, feed)
                 
-        except Exception as e:
-            logger.debug(f"Moving feed and checksum to failed directory, feed: {feed.name}, failed_dir: {self._failed_dir}")
+        except Exception:            
+            if not self._preserve_source_feeds: 
+                logger.debug(f"Moving feed and checksum to failed directory, feed: {feed.name}, failed_dir: {self._failed_dir}")
+
+                if feed and feed.exists():
+                    self._failed_dir.mkdir(parents=True, exist_ok=True)
+                    feed.rename(self._failed_dir / feed.name)
             
-            if feed and feed.exists():    
-                self._failed_dir.mkdir(parents=True, exist_ok=True)
-                feed.rename(self._failed_dir / feed.name)
-            
-            md5_file = feed.with_suffix(self._checksum_extension)
-            if md5_file and md5_file.exists():
-                md5_file.rename(self._failed_dir / md5_file.name)
+                md5_file = feed.with_suffix(self._checksum_extension)
+                if md5_file and md5_file.exists():
+                    md5_file.rename(self._failed_dir / md5_file.name)
                 
             if unpacked_dir and unpacked_dir.exists():
                 self._delete_path(unpacked_dir)
@@ -56,9 +61,13 @@ class Pipeline:
         
         try:
             logger.debug(f"Deleting started...")
-            self._delete_path(feed)
-            self._delete_path(feed.with_suffix(self._checksum_extension))
+            
+            if not self._preserve_source_feeds:
+                self._delete_path(feed)
+                self._delete_path(feed.with_suffix(self._checksum_extension))
+            
             self._delete_path(unpacked_dir)
+        
         except Exception as e:
             msg: str = f"Error deleting local files for feed: {feed.name}"
             logger.error(f"{msg}, error: {e}")
@@ -68,11 +77,17 @@ class Pipeline:
         logger.debug(f"Verifying checksum, feed: {feed.name}")
         expected_checksum_file: Path = feed.with_suffix(self._checksum_extension)
         expected_checksum: str
+        
+        if not expected_checksum_file.exists():
+            msg: str = f"Checksum file does not exist for feed, skipping feed. Feed: {feed.name}, expected: {expected_checksum_file}"
+            logger.warning(msg)
+            return
+        
         try:
             with open(expected_checksum_file, 'r') as f:
                 expected_checksum = f.read().strip()
         except Exception as e:
-            msg: str =  f"Checksum file missing: {expected_checksum_file}"
+            msg: str =  f"Checksum file cannot be open: {expected_checksum_file}"
             logger.error(f"{msg}, error: {e}")
             raise IOError(msg) from e
 
@@ -88,23 +103,25 @@ class Pipeline:
         
     def _unpack(self, feed: Path) -> Path:
         logger.debug(f"Unpacking feed: {feed.name}")
+
+        temp_dir: Path = (self._processing_dir / datetime.now().isoformat())
+        
         try: 
             with tarfile.open(feed, "r") as tar:
-                self._processing_dir.mkdir(parents=True, exist_ok=True)
-                temp_dir = tempfile.mkdtemp(dir=self._processing_dir)
+                
+                temp_dir.mkdir(parents=True, exist_ok=True)
                 logger.debug(f"Extracting {feed.name} to {temp_dir}")
                 tar.extractall(temp_dir, filter='data')
 
                 return Path(temp_dir)
         except Exception as e:
-            msg: str = f"Corrupted archive (feed), cannot extract {feed.name} to processing dir: {self._processing_dir}"
+            msg: str = f"Corrupted archive (feed), cannot extract {feed.name} to processing dir: {temp_dir}"
             logger.error(f"{msg}, error: {e}")
             raise IOError(msg) from e
     
     def _verify_feed_content(self, unpacked_dir: Path, feed: Path) -> None:
         control_file_ext: str = ".control"
         logger.debug(f"Verifying feed content, feed: {feed}, unpacked in: {unpacked_dir}")
-        logger.debug(f"DUPA 0")
         
         matching_file: str = next((f for f in os.listdir(unpacked_dir) if f.endswith(control_file_ext)), None)
 
@@ -117,6 +134,7 @@ class Pipeline:
         raise ValueError(msg)
     
     def _order_feed_content(self, unpacked_dir: Path, feed: Path) -> list[Path]:
+        logger.debug(f"Ordering feed content, feed: {feed}, unpacked in: {unpacked_dir}")
         file_list: list[str] = os.listdir(unpacked_dir)
         for i, e in enumerate(file_list):
             if e.endswith(".control"):
@@ -124,22 +142,24 @@ class Pipeline:
                 file_list[-1] = file_list[i]
                 file_list[i] = tmp
                 break
+
+        filtered_list = list(filter(lambda x: not x.startswith('._'), file_list))
         
-        if not file_list or not file_list[-1].endswith(".control"):
+        if not filtered_list or not filtered_list[-1].endswith(".control"):
             msg: str = f"Last file is not a *.control file in {unpacked_dir} for feed {feed}."
             logger.error(msg)
             raise ValueError(msg)
-        return [Path(f) for f in file_list]
+        
+        return [unpacked_dir / f for f in filtered_list]
 
     def _upload(self, ordered_feed_content: list[Path], feed: Path) -> None:
-        err_list: list[(Path, str)] = []
         for file in ordered_feed_content:
             try:
-                self._az_copy.upload(file)
+                self._az_copy.upload(file.absolute())
                 logger.debug(f"Upload succeeded for {file} from feed {feed.name}")
             except Exception as e:
-                msg: str = f"Upload failed for {file}, in feed: {feed}"
-                logger.error(f"{msg}, error: {e}")
+                msg: str = f"Upload failed for {file}, in feed: {feed}, underlying error: {e}"
+                logger.error(msg)
                 raise IOError(msg) from e
 
     @staticmethod
